requirements.txt
# Core dependencies
rich>=13.0.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0
asyncio>=3.4.3

# API providers
openai>=1.0.0        # OpenRouter, Groq compatibility
google-generativeai>=0.3.0  # Gemini
mistralai>=0.1.0     # Mistral AI

# Optional: TUI enhancements
textual>=0.40.0      # For advanced TUI (optional)

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
